more unit tests added
