+++ test/Selenium.WebDriver.Extensions.Core.Tests/Loaders/QuerySelectorLoaderTests.cs
@@ -9,14 +9,14 @@
     public class QuerySelectorLoaderTests
     {
         [Fact]
-        public void ShouldLoadScript()
+        public void ShouldReturnNullWhenLoadingQuerySelectorScript()
         {
             var loadScript = new QuerySelectorLoader().LoadScript();
             Assert.Null(loadScript);
         }
 
         [Fact]
-        public void ShouldGetDefaultLibraryUrl()
+        public void ShouldReturnNullWhenGettingQuerySelectorLibraryUrl()
         {
             var url = new QuerySelectorLoader().LibraryUri;
             Assert.Null(url);
